requirements.txt
numpy
torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0 # ?
torchtext==0.12.0 # ?
speechbrain==0.5.12 # ?
openai-whisper
pyannote-audio==2.1.1 # ==0.0.1
pywhispercpp
# sndfile